Fix clang warning.
BUG=None
TEST=None

Review URL: http://codereview.chromium.org/7301004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@91345 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0a33cf6ee4801aaf4753c568ed249c69f8a9d708

+++ ppapi/proxy/ppb_flash_tcp_socket_proxy.cc
@@ -298,7 +298,8 @@ void FlashTCPSocket::OnReadCompleted(bool succeeded, const std::string& data) {
 
   PP_RunAndClearCompletionCallback(
       &read_callback_,
-      succeeded ? data.size() : static_cast<int32_t>(PP_ERROR_FAILED));
+      succeeded ? static_cast<int32_t>(data.size()) :
+                  static_cast<int32_t>(PP_ERROR_FAILED));
 }
 
 void FlashTCPSocket::OnWriteCompleted(bool succeeded, int32_t bytes_written) {
